Add fzf plugin & loading the key-bindings loading from here
Previously, the configuration required loading the fzf key-bindings from
a separate file ~/.fzf.zsh with the OS specific path for the fzf files.
However, since I'm on Debian nearly all the time, it's not useful for me
to be quite that general with it. This adds the OS specific loading of
the key-bindings to this repo so there's no extra setup.